requirements.txt
asttokens==2.2.1
backcall==0.2.0
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
deepevolution==0.0.4
diff-match-patch-python==1.0.3
executing==1.2.0
fonttools==4.38.0
func-timeout==4.3.5
geneticalgorithm==1.0.2
imageio==2.26.0
install==1.3.5
ipython==8.11.0
jedi==0.18.2
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.1
lxml==4.9.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==3.0
numpy==1.23.5
packaging==21.3
pandas==1.5.1
parso==0.8.3
pdf-diff==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
stack-data==0.6.2
tifffile==2023.2.28
tqdm==4.64.1
traitlets==5.9.0
wcwidth==0.2.6